Fixed casting of dPVal for genfit convergence

+++ Libraries/TAGfoot/TAGactKFitter.hxx
@@ -108,7 +108,7 @@ public:
 
 private:
 
-	AbsKalmanFitter*	InitializeFitter(int nIter, int dPVal);
+	AbsKalmanFitter*	InitializeFitter(int nIter, double dPVal);
 	TAGFselectorBase*	InitializeSelector();
 	int		MakeFit(long evNum, TAGFselectorBase* selector);
